another cli refactor
